FACE_DETECTION/retinaface/CephTest/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
PyYAML>=6.0
tabulate>=0.9.0
requests>=2.28.0
